+++ sys/kernel/mem.h
@@ -1,9 +1,11 @@
 #ifndef __MEM_H
 #define __MEM_H
 
+#include <arch/types.h>
+
 /* Frees all machine memory regions mapped as usable. Returns 0 normally and -1
    in case of error. */
-int freeallmem(void);
+u8 freeallmem(void);
 
 /* Disables virtual memory to use physical instead. */
 void physicalmem(void);
